[ledger] Switch use of Cobalt to new Cobalt 1.0 project.
Also use NewCobaltLoggerFromProjectName() instead of NewCobaltLogger().
This means that when connecting to Cobalt, rather than passing a VMO
containing a serialized protobuf containing our metric definitions
we instead assume that the version of the metrics registry that is
bundled with Cobalt contains the latest version of our metric definitions
so instead we only pass the name of our project.

Also we take advantage of Cobalt's ability to generate C++ source code
containing constants for the metric IDs.

Test:
I ran all of the Ledger tests.

Change-Id: Ie4c70f82077d10dbd4eecf9756be6670d6d50592