[media] Move audio examples into new source layout
This CL moves the media example source code out of garnet/examples/media
and topaz/examples/media into examples/media/audio. It also relocates
the text-to-speech example to examples/speech/tts, and renames it to
'speak' (to avoid name collision with the overall tts area). Finally,
it renames simple_sine_sync to simple_sync to avoid name collision
when executing 'run simple_sine' from the fx shell.

This is all in accordance with the previously-agreed-upon source
layout plan. There should be no behavior change to product or tests.

Test: no behavior change

MTWN-278 #Done

Change-Id: Ibd4a9c96d2d85fab52448c1ed7c92dc3a7e8c102