[docs][fidl] Start of Bindings Specification
Following FTP-024, we decided to create a Bindings Specification page in
order to heighten the rigor and expectations on FIDL bindings. Work is
ongoing to translate the content of FTP-024, and this is a start to
provide overall structure.

Additionally, capturing a recent conversation about recommendations
around bits support in bindings.

Test: read the docs
Change-Id: Ib94a676d67ffdcdaa3db42540923feca213613f4