[feedback] actually forward the dispatcher to the bindings in the stub
the behavior is technically unchanged as in practice the stub was still
using the dispatcher of the second loop as the default dispatcher will
resolve to that one by the time the ServiceDirectoryProvider serves the
stub. Still better to be explicit.

TESTED=`fx run-test feedback_agent_tests -t log_listener_unittest`

Change-Id: I1dcad6d677652a182df87776180a88bed43d18d0