[fidl] Constrain ImplPtr to various pointer types
We are looking to add functionality which requires equality semantics on
ImplPtr. This statically asserts that we are only using, or allowing
use, with plain vanilla pointers, unique pointers, or shared pointers.

We are choosing a restrictive allow-list based static assert to overly
constrain uses today. The set of supported ImplPtr types can be enlarged
later. This list was put together to be minimal, given uses in our
codebase today.

FIDL-344 #comment

Test: builds
Change-Id: I74d26e617e17f3238ac4170242750ac90fc3a932