[zircon][zxtest]: Macros for adding failures.
ADD_FAILURE -> Adds non fatal failure and does not return.
ADD_FATAL_FAILURE -> Adds fatal failure and does not return.
ASSERT_NO_FAILURES -> ASSERTS if any failure ocurred and returns.

In order to make this cleanly, reworked assertion a little bit. This
will make this manual failures display cleaner errors instead of
exposing the value of internal macros(_ZXTEST_*).

Added a new attribute to assertion, which marks it as a value based
assertion or a manual assertion. This allows the reporter not to print
extra information when this happens, and have cleaner error messages.

ZX-4008 #done

TEST=zxtest-integration-test

Change-Id: I1ec736fba7d5ac4e8b2175af92f51efa3c7fad26