[debugger] add registers to StackFrame.
Adds a register vector to the stack frame record that is sent from the
debug agent to the client. This will be used in the future to send known
register values for each stack frame along with the backtrace. Currently
the register values are unused.

Add more constructors and comparison functions to StackFrame and
Register to help testing. Update some tests to use these new functions
to improve readability.

Change-Id: I7c25f3fabdd153ff36877d9ddc26a59a2db5ad53