[appmgr] Statically link C++ std lib, remove fostr dependency
This removes the dynamic library dependencies on the C++ standard
library and the fostr library from appmgr to partially resolve
dependency cycles when booting up the system.  With this change, the
only shared library dependency remaining outside of the bootfs is
libfxl_logging.so, which stores a static log level.

By itself this patch has no behavior change in the system.

DNO-492 #comment remove most shared library deps from appmgr

Change-Id: I31394945f62e3ccff4bf61e80daa41b139913e7c