[blobfs] Replace journal info Buffer with VmoBuffer
"Buffer" provides a circular buffer interface.

It is being replaced by both:
- RingBuffer (circular buffer with safer semantics) and
- VmoBuffer (non-circular buffer).

This patch updates a usage in the journal of the info
block, where a Buffer was being used in a non-allocating
manner. This new interface is smaller, and harder
to misuse. This makes the call to "ReserveIndex" extraneous,
since allocation is simply disallowed by the type system.

This patch replaces an untested dependency of the journal
with a tested dependency. It otherwise has no observable
behavior.

Change-Id: Id169c13ba52a2b8d7705316c17d87f57760d6af4