[blobfs] Create BufferedOperation type
This allows clients to easily distinguish between
an "UnbufferedOperation" (not yet in a buffer registered
with the block device) and a "BufferedOperation", which
can directly be transmitted to storage via "FlushWriteRequests".

This change makes each operation self-describing, which makes
it easier to operate on collections of "BufferedOperation" objects
that might be acting on distinct objects (e.g., the Journal buffer
vs the data writeback buffer).

Change-Id: I98811f6e3a256f79fe83bd7e2ec13b717a38c3c5