[fshost] Reformat zxcrypt and minfs partitions when missing.
If the disk format is unknown but the GPT GUID values dictate that it
should be a data format, it determines if the volume should be a zxcrypt
or minfs partition and formats it as such.

Added a unit test that failed without the changes. I added a test that
would have caught the vim2 breakage, but I also manually tested that it
boots up properly.

Change-Id: I276d55fc8dbcc8377d6a6b7d5dfe7b0fd9019eb4