[debugger] Set registers in StackFrame
Adds all general registers to the topmost StackFrame when sending frame
information. These are currently unused.

Changes existing register handling code to reduce one query of the register
state when unwinding with the Android unwinder, and to re-use the register
saving code in "arch.h" without re-querying the OS.

Adds a unit test for stack unwinding which was previuosly missing. It
validates that the new registers are set, but not what the specific
values are (these are hard to predict).

Disables an existing test that we found to be flaky.

Change-Id: I9a873b04dfdf91b0bd4a5f1f96301ee1ef275dcd