[utest] Don't hard-code /boot location for zircon test files
This change fixes the zircon tests that were failing in the `bringup`
build.

The tests were failing because they had hardcoded paths at "/boot/..."
which failed when the tests were run from "/system". The fix is for
runtests to pass a TEST_ROOT_DIR environment variable, which is
"/system" or "/boot" depending on the location of the test. The tests
use this variable instead of hardcoding the path.

Bug: BLD-463 #comment fix broken tests
Change-Id: Id6478f68600a55f13dc8c0bde19253c909a03694