[kernel][vm] Add bidirectional COW clones
Bidirectional COW clones are implemented by moving the original vmo's
content into a kernel-private 'hidden' vmo and using that hidden vmo as
the parent of the original vmo as well as the new bidirectional COW
clone.

The hidden vmos are an implementation detail of bidirectional clones,
and are thus not exposed to userspace through any syscalls. However,
they do show up in 'k' commands, as those commands are not a stable API.

Subsequent changes will improve the memory consumption of bidirectional
clones.

ZX-1268 #comment [kernel][vm] Add bidirectional COW clones

Change-Id: I4a7f46ff92db2746d1f540c1ca84aae6281edb5b