[scenic] Opacity support is a HACK.
(this relands I5705b27f46002b904e8d025e360b30f501871de7, modified
for a soft transition)

For reasons outside the scope of this CL description, Scenic
wishes to avoid the use of translucent objects within 3D scenes;
modifying opacity should only be available as a 2D operation.
However, Scenic does not yet support 2D scenes; "everything" is
currently part of a 3D scene.  As a short term solution, we
therefore allow OpacityNode to modify the opacity of shapes in
the 3D scene, but mark it as a HACK to indicate that this support
will be removed at some point in the future.

SCN-478 #comment

Testability: no functional change.

Change-Id: Iec09f956655ca1fa49c241066d70b7b5dabd948b