[media][tests] Add GetNumDevices to virtualaudio
This test-only CL adds a GetNumDevices method to the
VirtualAudioControl protocol. This will be used by test
clients (e.g. AudioDeviceTests), to synchronize with the
Disabling of virtualaudio (which removes all active virtual
audio input and output devices) or the activation of a
virtual audio device. Along with this, the virtual_audio
CL utility is updated to interactively exercise this new
test function.

VirtualAudio infrastructure is used by AudioDeviceTests (to
validate the AudioDeviceEnumerator and AudioCore services)
and by the E2E tests. Unfortunately, AudioDeviceTests is
currently disabled because of synchronization issues, which
poses a testability chicken-and-egg for this CL. Once this
CL lands, the changes to re-enable AudioDeviceTests are
already prepared and will come immediately after.

Test: virtual_audio driver and virtual_audio utility are test-
only, as are the fuchsia.virtualaudio FIDL protocols.

MTWN-293 #comment

Change-Id: I0c3f7b0560df7b7e0d103d5b310fc6e86e189551