[media][tests] Move AudioLoopbackTests to use virtualaudio
With this CL, AudioLoopbackTests now instantiates a virtual
audio output device, so that loopback capture is always
usable. To wait reliably for the device to be present (and
to wait reliably for it to be gone before exiting), it must
use GetNumVirtualDevices in the same way that
audio_device_tests does.

This test-only CL has no effect on production code

Test: build, CQ

Change-Id: I19bada2c69a0ea4fba7c05317ceaaa7355a9c1b4