[fuchsia] Commit initial starlark and regenerate
This is the first iteration of starlark code used to
generate the protobuf APIs for CI/CQ builds. The starlark
is basically a 1:1 mapping for now. In subsequent changes
we'll pull common code out into reusable libraries.

infra is still bootstrapping this infrastructure, so there's
not much documentation yet. Soon to come.

Bug: IN-1198 #comment
Bug: IN-1102 #comment

Change-Id: Ib994df041388b98ddeb520aada8d3cca89b3ea56