[buildidtool] Update to handle groups of DSOs
This changes alters the way the tool works in order to be more efficent.
It handles groups of DSOs that all have the same build ID. Later we might
add support for additional files like .dwp or .gsym files which won't have
build IDs at all but still should be associated. We will see.

Change-Id: I22a433b0910c9b0f6ff945f1b30cefc3c98a0799