[debugger] Remove BP from StackFrame.
The BP member of the StackFrame originally came from the unwinder but
normally we use the frame base from the symbols since the BP register
isn't necessarily used (it never is for our release builds). In the
current default mode (Android unwinder), it's always set the same as the
stack pointer so is actively misleading.

Removes printing the BP in the long backtrace format. This doesn't
handle async results and now sometimes the base pointer requires
asynchronous evaluation. We can add this back later if we desire.

Other than the printing change, there should be no noticeable behavior
change in normal usage so the existing tests should cover it. IN some
"symbols corrupt or missing" cases, the default BP value changed from
the register to 0, but 0 is probably less misleading when the base
pointer isn't really known.

Change-Id: Id6a5f18d27186bc2337957dfefda5fc26ca21cb3