[audio_core] Admin method to control device settings
This CL adds an admin-only method to enable or disable the
systemwide creation and update of audio device settings files.

This work is being done in preparation for the creation of long-
running audio tests that (among other things) create and destroy
an endless number of audio input and output devices (in conjunction
with virtualaudio), which otherwise would completely fill all
available storage with audio device settings files.  While we do
also want eventually to test the out-of-storage scenario, we also
want to be able to remove this factor, in order to find the other
factors that may stop us from being able to operate indefinitely.

Test: to be created

Change-Id: I1f0f9fff815b6c6f9d088c50b76d961f852902bb