[minfs] disk-inspector Part 2: Implement the minfs inspector backend for on-disk datastructures.
This change implements the minfs inspector backend for different on-disk structures like
superblock, inode table, inode, journal et al and returns each of these as a generic DiskObject.
Please note that alloc block bitmap and inode bitmap are not implemented as part of this
change and will be implemented as a later CL.
This CL also splits the Minfs class and InodeManager class to make them more easy to accomodate
unit testing.

Testing: Added unit tests for root, superblock, inode table, inode and journal.

Change-Id: I3bed793a8c0301bf8ce4367fbca48e31a6f44772