[component_manager] Use log crate's macros instead of defining our own
This integrates the component_manager logger, which sends logs to the
kernel debuglog facility, with the log crate instead of defining our own
macros. It also adds a panic hook to ensure that panics are also sent to
the debuglog, rather than relying on component_manager's stderr to be
wired up properly.

Also removes the stderr fallback path, which is unnecessary.

Change-Id: I3510b04239c7f9ab306254663bddaab245f48197