[zircon][gn] Fix environment_redirect() for special-case selectors lists
When an environment() sets variant_selectors so it doesn't match the
default list used in common environments, things get even hairier for
environment_redirect().  Previously environment() documented that
default_variants must be in the list for environment_redirect() to work
right, but even that was not actually enough.  Instead, remove that
requirement and make environment() always define a dummy environment with
the "default redirect" variant name that environment_redirect() will
choose.  Make that logic choose only from default_variants and line up with
the fallback dummy variant name logic in environment().

Bug: BLD-430
Change-Id: I399a38ba494f4fafc28a3d9e93dece101eab3732