[fidlc] Unique-ify nullable types in coding tables
Prior to this, if you had multiple references to a nullable xunion in
the flat AST, it would include multiple copies of that xunion in the
list of coded types.  Since the list of coded types is directly used to
generate xunion declarations (see TablesGenerator::Produce), it could
emit multiple versions of the same nullable xunion into the C generated
code, leading to compilation failure.

Also applies the same logic to nullable structs and unions, although
they don't have the same failure mode.

Change-Id: I840beaa607f390a9ec7e685d6c0e37477445d657