[brcmfmac][wlan] Add simulator configuration
Add a simulator configuration to lay the groundwork for driver
unit tests. This first step creates a build target that doesn't
use the zircon driver/device framework. To do so, we create two
sets of sources, one (for the actual driver) that instantiates
a devmgr device and one (for testing) that instantiates a
simulated device.

WLAN-1126

TEST: Built, ran brcmfmac unit tests, ran smoke test with
      fx-linksys. This patch is the first step in a unit test
      framework for the brcmfmac driver.
Change-Id: Icecadcfd75f727084447953c6a521775ca499737