[fidl][cpp] Introduce fidl::EventSender
The fidl::EventSender is similar to fidl::Binding except the
fidl::EventSender just sends events and does not receive or decode
incoming messages.

Having a separate class for this functionality is useful because
EventSender does not need to couple to an async_dispatcher_t, which
means EventSender can be used in scenarios where an async_dispatcher_t
is not available or when the client wishes to send events from multiple
threads.

This CL also uses EventSender in lib/vfs/cpp in a location where an
async_dispatcher_t is not available. This change lets us remove a
dependency on the C bindings in lib/vfs/cpp, which is the last blocker
for adding lib/vfs/cpp to the SDK. The change also lets us remove one
more instance of manually constructing FIDL messages.

Test: Added event_sender_unittest.cc
Change-Id: If9d6c9cca4d16d248d3635be677d35b61d321e28