[media][test] Test 3.999:1 and 1:3.999 resampling
This CL changes the previous new test cases (which validated the
3:1 resampling ratio) to instead validate the 3.9999:1 resampling
ratio, and to add the 1:3.9999 resampling ratio.  This tests
particularly problematic ratios for both interpolators, and continues
to defend our sample-rate limits (this effectively rate-converts from
191999 Hz to 48000 Hz).

This CL also includes fidelity limit values that we require for these
ratios (both frequency response and signal-to-noise-and-distortion).

Test: test-only CL; adds 8 test cases runnable on any product/board

Change-Id: I4d2ff29a48403eaa98cd13f721b66e38191211bc