[audio_core] Resampler cleanup
This CL aligns the Linear and Sample-and-Hold interpolators, in
preparation of introducing additional interpolating resamplers.
No functional change -- just tweaking the linear and point resamplers
to use identical patterns and sequences (as the soon-to-be-added
ones will likely do, as well).

Test: 170 existing automated test cases in audio_mixer_tests
(audio_mixer_tests --full) on all products.boards

MTWN-296 #done

Change-Id: I0753b3fc13d4eb69ffef56a38906043980304df1