[ISP] Test for ARM ISP
Added a test class to the ISP.  This class is instantiated
by the ArmIspDevice to offer a testing fidl interface when enabled.
Currently this class is always created, a future CL will allow
enabling the test interface with flags.

Also added the fidl interface definition for the test, as well as
a test program.

Test = This is the test.  Running this test!

Change-Id: I87c3568ae82cba4321bb4826ca0dacf016e51f0d