[magma] add magma json definition
This change adds a json definition for the magma ABI, and a script to
generate magma.h from it. It also updates magma.h to use the generated
output format. Future revisions to the magma ABI should be made in the
json file, and the script used to regenerate the corresponding header.
The script may be run as part of the build, but is currently only done
so by a new codegen test path in the magma test suite.

Test: used pycparser to verify magma.h AST unchanged from current version
      changed magma.json and verified build failure
      changed magma.h and verified build failure
      changed both magma.json and magma.h and verified build success
Change-Id: Id55f0d28ec89adbcaba3b53ed7647852d9a84127