[fuchsia-zircon] Create Rust wrappers for zx_debuglog_* syscalls
And use them in places where we make such syscalls directly with unsafe
Rust today. (Plus add in missing test coverage for logger that covers
the modified code.)

The zx_debuglog_read wrapper currently returns a plain Vec of bytes and
the caller needs to parse it according to the definition of
zx_log_record_t. In the future it ideally would return an already-parsed
Rust struct type, but there is some uncertainty around the existing
debuglog syscalls so this wasn't done yet.

Change-Id: Ibf3ddc6195a751dbacb6e0e9d24250370e8516e9