[debugger] Remove stack address checking in test.
This test of code addresses is not flexible enought to account for all
the code our compilers can generate. Testing this in the future will
require something more complicated like assembly code, so this test
removes that part of the checking for now. The rest of the test is still
a good smoke test for the unwinder.

Change-Id: Id2d18eca6ccb8b4d214a3bf627e6da090388542f