[wlan-inspect] Be explicit about logging Option
Previously, we added a blanket WriteInspect implementation to Option
types that would log a value only if an Option doesn't exist. When
looking at an inspect_log call on an optional value, it's non-obvious
that in some cases, no value would be logged, and this blanket
implementation was only added because inspect_log didn't have a way to
specify optional logging.

With this commit, optional logging can be specified with a question
mark on the key, so the blanket WriteInspect implementation is also
removed.

Test: fx run-test wlan-inspect-tests
Change-Id: I926c20fe2e64fc2e1221c2b84b61d5d9b062e817