[fdio][llcpp] Start of migration to llcpp
Begin in remoteio.cpp, by rewriting calls to the C binding as llcpp.

Public APIs converted:
- fdio_open_at
- fdio_service_connect_at
- fdio_service_clone
- fdio_service_clone_to

As of this patch, some implementations are less than ideal: most
helper functions take zx_handle_t as arguments as opposed to zx::channel
or zx::unowned_channel, which encodes ownership in the type. As a
result, there are frequent casting from zx_handle_t to zx::channel etc.
and vice versa. As greater proportion of the library is converted to
C++, these conversions may be removed.

Change-Id: I9a8bc4df228610daafa64eec7ebc99ba194a3489