[zircon][gn][libc] Add ubsan-sancov-full variant
This new variant enables all the kinds of sancov instrumentation,
not just the basic one that the standard runtime handles.  This
exercises more of the instrumentation features that e.g. fuzzing
also uses.  The special magic in the dynamic linker to cope with
the instrumentation is extended to handle all the cases.

Bug: ZX-3986 #comment ubsan-sancov-full variant
Test: /boot/test/sys/noop-test.ubsan-sancov-full
Change-Id: I2ca2bbaca732526ee10e0d60822410f48d8eb1ef