[bootsvc] Add fuchsia.boot.RootJob service
This allows processes further down the job-hierarchy to be able to get
access to the RootJob. This is useful for processes such as the
devcoordinator (once component_manager is in place) and utilities like
ps.

Eventually, we will want to further limit or entirely get rid of this
interface. That will happen once component_manager is in place.

ZX-4002
ZX-4072

Change-Id: I4b1e75a7841a04ea2cc7da374e42257d4802e619