[minfs] disk-inspector Part 3: CmdLine/Uapp application to consume the disk-interface
This change implements a very simple command line tool/uapp application to consume the
disk objects for minfs and recursively iterate over all of them.
Please note that this is not interactive at the time being, but is implemented
as a first implementation of the disk-inspector tool. It is a good way to
test the different apis of the disk-inpector library.

Testing: The application itself is pretty dumb and doesn't need unit tests.
          The unit tests for minfs backend are already added as part of the change:
          https://fuchsia-review.googlesource.com/c/fuchsia/+/277175
Change-Id: Icd211c30bdb2c82c9f242447d8fddc5d7620c323