[banjo] Support more types as union fields
This causes union fields to be generated using the same logic as struct
fields, so that arrays and strings are properly handled. Note for now
that vector fields are excluded from unions, as they actually generate
two fields (a pointer and a count) which does not interact well with
unions.

Change-Id: I69a36903390740d5443f5e3ff0d00d3e49e423f8