[kernel][x86] X86ArchVmAspace: Add basic unit test of one mapping
Add a basic unit test that creates a 4 GB X86ArchVmAspace, maps one page,
and examines the page tables to ensure the mapping is found and extraneous
mappings are not found.

This is a skeleton that will allow us to write unit tests for a lot more
exciting/complex page table manipulation / address space operations.

Change-Id: Ifb0add1fff871834b877ed3bf6e3121577450ad5