[devfs][test] Fix incorrect assumption in test
We can't assume that there will always be a block device. For now, we
can assume that there will always be a platform-bus driver bound.

It would be better if we created a fake devfs for these tests but since
device discovery is going to signficantly change in the near future this
should be good enough for now.

ZX-4090 #done

Change-Id: I668e6a13ec273f83d13654ba692ea92763fc41b8