[build] Fix Rust static library dependecies
Depending on a static_library from Rust fails because the correct
link directory path is not passed to build_rustc_target.py. This
fix adds link paths from every one of the dependecies found in
non_rust_deps.

This is not ideal since it doesn't work with source_set or any
shared library that has transitive dependencies.

TEST=Build, CQ, and Rust crate that tests scenario.
     fx run-test static_library_tests

Change-Id: I42c2d07e68d5f7a3d76c0edacbcd0d81f50c4993