[kernel][vm] Flatten unnecessary hidden vmos
If one of the children of a hidden vmo is destroyed, then the hidden vmo
can be merged into its other child. This prevents clone operations from
having a permanent cost of kernel metadata, and it provides a more
uniform child tree by removing 'useless' hidden vmos and ensuring that
every hidden vmo has exactly two children.

As children are removed in the vmo destructor, flattening vmos cannot be
able to fail. In particular, the page lists need to be able to be merged
without requiring any new allocations. To accomplish this, this change
adds a new skew field to VmPageList. This field is used to align all the
page lists in a child tree so that the VmPageListNodes can be directly
moved between page lists without being split.

ZX-1268 #comment [kernel][vm] Flatten unnecessary hidden vmos

Change-Id: I59b645106b80e880f3e014e5127c06bb738180e0