[build] Include tests manifest in aux manifest
When individual binaries are drawn from the tests manifest
via zircon_extras_manifest(), their libraries must be found
in the aux manifest even though they are also in the tests
manifest alongside the binaries.  This matters for binaries
with extra variants like fuzzers, since only the vanilla and
asan libraries are exhaustively included in the aux manifest.

Bug: BLD-430
Change-Id: Id8f1abe29d74832a612bcfe064b41ff278ef5e73