[escher] Support for 3D clip planes in PaperShapeCache::GetShapeMesh()
Allow for any arbitrary 3D clipping plane to be projected onto the
z=0 plane, resulting in an edge line that can be used for clipping
2D meshes that live on that plane. Previously there was a FXL_DCHECK
to ensure that the z component of the incoming 3D clip plane equaled
zero, and this change is meant to remove that restriction.

In the case of a 3D clipping plane that is parallel to the z=0 plane,
we now simply skip it.

Also adds in a small epsilon buffer to bounding boxes when testing
to see if they are being clipped by a plane. The purpose of this is
to make clipping slightly less aggressive in order to handle cases
such as when an object is resting directly on the floor of the
bounding box. These objects should be treated as existing within
their allotted space and should not be clipped.

ES-141 #done

This is a necessary feature to finish before SCN-1180 (Enforce
Clipping on Z-Bounds of a View - which is on the list of features to
complete for our testing effort), can be done, as ensuring that the
contents of a view are properly clipped by its bounding box requires
that escher can handle clipping arbitrarily aligned clipping planes.

Change-Id: I7a248e5f8f101bbc49cff5b0c458f044d96e526f