[modular] Fix crash if TestHarness config is too big
Found problem in pseudo_dir_utils.cc that creates a PseudoFile for the
modular config generated by TestHarness. If the data generates a config
longer than 1024 bytes, it crashes. This change sets the max size to the
size of the config string, resolving the problem.

Added new unittest.

See fxb/MF-411

Test: fx run-test peridot_tests

Change-Id: Ie2e9e9b25883f6ccb55b379ac307200c174e2e56