[kernel][threads] fix thread test synchronization
We've been seeing some flake on theads-test, it's been difficult to
reproduce locally so far but the test is improperly using "volatile" to
synchronize a variable between threads and it seems like this could
potentially be causing the issue.

Additionally, this adds some further analysis into the failing function
so that if we continue to see flake we'll have some more data to work
with.

FLK-233

Test: runtests
Change-Id: Ia1314fa1363663cfd65b6836a33f513dd637a0d7