[zircon] Make vmos non-resizable by default
This change makes vmos non-resizable by default and adds the resizable
flag to all necessary locations in the fuchsia source tree. The flag has
already been added to compiler-rt, which is the only out-of-tree caller
of zx_vmo_set_size.

To maintain source-level compatibility with out-of-tree clients, the
non-resizable flags are redefined as 0. To maintain binary
compatibility, the kernel still acknowledges the old flag values.

ZX-2357 #comment [zircon] Make vmos non-resizable by default

Change-Id: If75fe43e81abee5dca3b26cc170d538f373ff594