[fidlcat] Two step message decoding
This is needed to be able to do a pretty print.

A first step decodes a message and generates an object.

A second step uses this object and generates the JSON.

There is no new tests because this cl doesn't add any
functionality.

fx run-test fidlcat_tests
[==========] 58 tests from 2 test suites ran. (113 ms total)
[  PASSED  ] 58 tests.

DX-1482 #comment Two step message decoding

Change-Id: Id0fcf4c5921465e7b689ee0267879b403842ee20