[build] Support "with_shared" properly in non-Fuchsia toolchains
The with_shared=true parameter was already the default in
clang_toolchain_suite() so passing it for the linux toolchains was
superfluous.  But the plumbing that makes this actually work was
gated on is_fuchsia.  Instead, plumb through the with_shared flag
via toolchain_variant and test that in BUILDCONFIG.gn.

This makes it possible to use shared_library() targets correctly in
the linux toolchains.

Bug: BLD-475 #done
Change-Id: Idda02f28486c1b0e03e407ec7f52ac898ea609ee