[fzl] Fix race in memory-probe
After observing a crash and killing the probe thread, wait for the
thread to actually terminate before returning from the function and
closing the exception port. This fixes a race where closing the port can
forward the exception to the next handler even if zx_task_kill has
already returned (as that syscall is defined to be asynchronous).

FLK-238 #done [fzl] Fix race in memory-probe

Change-Id: I3c8f814efeae7f593df3d99d05346570982f0895