[kernel][tests] mmu_tests: Move common test cases from x64 -> common
Zircon had a small number of tests of the ArchVmAspace interface in the
arch/x86 subtree; move them to common code so that they run against both
Arm64 and x86.

This commit does not change the tests except as needed to allow them to
build/run on both platforms. A subsequent CL will split them into subtests
and perform other cleanups suggested in
https://fuchsia-review.googlesource.com/c/fuchsia/+/281509

Change-Id: Ifd7b5c68d09e29749e0fd01363eb1d9eca102dec