[scripts][fuzzing] Fix libFuzzer directory options
It turns out libFuzzer is very literal with options like
'artifact_prefix': it directly prepends the value given to test
artifacts, and make no attempt to distinguish when the prefix is a
directory.  This CL adds a '/' to the end of the artifact prefix to get
the test artifacts written to the output directory, as expected.

Bug: SEC-287
Test: fx set core.x64 --with //src/fuzzing:tests
      fx run-host-tests fuzzing_host_test
Change-Id: I33674f0affdf8fa04b828108ef04f431d5c2520f