[identity] Use buffered reading and writing for tokendb.
Integration tests have been taking a very long time (tens of seconds)
on the less powerful debug target and troubleshooting shows a large
chunk of this time is spent in serialization, since our file system is
not well suited to a large number of small reads and writes.

Wrapping the files in buffer types substantially improves performance.

Change-Id: I5ea7462235167d15f3a2cd598ef384cbcc7b9428