[rust][fuchsia-component] Add Vmofile support to ServiceFs
This is a re-land of d485d3b6, which was reverted due to a
panic that could be caused by attempting to connect to a
service node using the NODE_REFERENCE flag, which would cause
serve_connection_at to be called on a service node, which
was previously invalid. This updated change adds support for
serving NodeRequest-only clients, allowing NODE_REFERENCE
requests for service nodes to be served correctly.

This is needed to unblock the use of the Inspect API in Rust,
which is a critical tool for testing Rust components.

Test: vmofile test
Change-Id: I2dd8beec613964b2fc55fcb397743e913d5fe024