[test][logger] Add FX_PLOGS() that logs zx_status_t as string
And also FX_VPLOGS. Similar to FXL_PLOG() recent addition, this suffixes
the message with "%d (%s)" for the zx_status_t given.

This does not handle the third style, FX_LOG() yet. It seems less-used
and more complicated to handle so it might not be worth it.

ZX-4110 #comment [test][logger] Add FX_PLOGS() that logs zx_status_t as string

Test: new tests in logger_unittest.cc
Change-Id: I7c65dbb6b110f3c0b9b785b1a1de7e5171382d26