[modular] Automatically start running a newly started story in StoryProvider.
Updates relevant integration tests to expect the stories to start
automatically.

Also updates the TestSessionShell (part of the modular
test harness) to not automatically start stories, since that's done
anyway now (note that this change leaves the OnUpdate() function there in case
it's needed in the future for something like recording story state
changes in tests).

TEST=fx run-test modular_tests

MF-271 #done

Change-Id: I0fb8a9def8b9a385fa3420e21193ae0685040a46