[sshutil] Add a total timeout to SSH client creation
ssh.Dial can hang during authentication, the 'timeout' being set in the config only applying to establishment of the initial connection. This function is effectively ssh.Dial with the ability to set a deadline on the underlying connection.

See https://github.com/golang/go/issues/21941 for more details on the hang.

Change-Id: Ie5b51dad23ac44bed5f98c322a5da8c06cd93f45