[mtk-power] Implement RequestVoltage
This change does the needful to request a particular voltage for a
regulator. The caller will be given the actual output voltage that has
been set. It is either the immediate step voltage or supported voltage that is
less than the requested voltage. If the caller requested a voltage that
is not within the supported range, error is returned.

Test: runtests -t mtk-power-test. Modified the display driver like so
https://fuchsia-review.googlesource.com/c/fuchsia/+/280469 to use the
power protocol and verified display comes up as before.

Change-Id: I7a88da48d75586e3f040e6a519353f6c71d9d6f2